requirements.txt
amqp==5.1.1
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-import-export==3.2.0
et-xmlfile==1.1.0
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
kombu==5.2.4
MarkupPy==1.14
numpy==1.24.4
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
prompt-toolkit==3.0.38
psycopg2-binary==2.8.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tablib==3.4.0
tornado==6.3.1
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
